Drop secret_service in favor of keyctl; setup for providers
