+++ Cargo.toml
@@ -14,7 +14,6 @@ napi-derive = "2.9.1"
 
 [features]
 default = []
-keyctl = ["dep:nc"]
 
 [target.'cfg(target_os = "windows")'.dependencies.windows]
 features = [
@@ -27,8 +26,7 @@ version = "0.43.0"
 security-framework = "2.7.0"
 
 [target.'cfg(any(target_os = "freebsd", target_os = "linux"))'.dependencies]
-nc = { version = "0.8", optional = true }
-secret-service = "2.0.2"
+nc = { version = "0.8" }
 
 [build-dependencies]
 napi-build = "2.0.1"
