+++ package.json
@@ -29,7 +29,6 @@
     "artifacts": "napi artifacts",
     "build": "napi build --platform --release",
     "build:debug": "napi build --platform",
-    "build:headless": "napi build --features keyctl --platform --release",
     "prepublishOnly": "napi prepublish -t npm",
     "test": "ava",
     "version": "napi version"
