+++ src/providers/mod.rs
@@ -1,1 +1,2 @@
+#[cfg(all(any(target_os = "freebsd", target_os = "linux")))]
 pub mod keyctl;
